requirements.txt
PyGObject>=3.42.0
mutagen>=1.46.0
Pillow>=9.0.0
PyPDF2>=3.0.0